# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Budget Guru API Swagger Documentation",
        "description": "Test",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "http://cop4331test.herokuapp.com",
    "basePath": "",
...
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json",
        "x-auth-token"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/api/dashboard/updateSpent": {
            "post": {
                "tags": [
                    "Dashboard Actions"
                ],
                "summary": "Adds value in spent container for the account",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "New spent amount and category",
                        "schema": {
                            "properties": {
                                "Spent": "number",
                                "Category": "string"
                            }
                        }
                    },
                    {
                        "in": "header",
                        "name": "x-auth-token",
                        "required": true,
                        "description": "Authorization token",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Spent amount added",
                        "schema": {
                            "properties": {
                                "msg": {"type": "string"},
                                "account": {
                                    "properties": {
                                        "Active": {"type": "boolean"},
                                        "Budget": {"type": "number"},
                                        "Earned": {"type": "number"},
                                        "Spent": {"type": "number"},
                                        "MonthlyIncome": {"type": "number"},
                                        "MonthlyBill": {"type": "number"},
                                        "CalculatedCategory": {
                                            "properties": {
                                                "Clothing": {"type": "number"},
                                                "FoodDrink": {"type": "number"},
                                                "Home": {"type": "number"},
                                                "Entertainment": {"type": "number"},
                                                "Transportation": {"type": "number"},
                                                "Health": {"type": "number"},
                                                "Misc": {"type": "number"}
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to add spent amount",
                        "schema": {
                            "properties": {
                                "error": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/dashboard/updateEarned": {
            "post": {
                "tags": [
                    "Dashboard Actions"
                ],
                "summary": "Earned amount added to budget",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "New earned amount",
                        "schema": {
                            "properties": {
                                "Earned": {"type": "number"}
                            }
                        }
                    },
                    {
                        "in": "header",
                        "name": "x-auth-token",
                        "required": true,
                        "description": "Authorization token",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Earned amount added",
                        "schema": {
                            "properties": {
                                "msg": {"type": "string"},
                                "account": {
                                    "properties": {
                                        "Active": {"type": "boolean"},
                                        "Budget": {"type": "number"},
                                        "Earned": {"type": "number"},
                                        "Spent": {"type": "number"},
                                        "MonthlyIncome": {"type": "number"},
                                        "MonthlyBill": {"type": "number"},
                                        "CalculatedCategory": {
                                            "properties": {
                                                "Clothing": {"type": "number"},
                                                "FoodDrink": {"type": "number"},
                                                "Home": {"type": "number"},
                                                "Entertainment": {"type": "number"},
                                                "Transportation": {"type": "number"},
                                                "Health": {"type": "number"},
                                                "Misc": {"type": "number"}
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Unable to add earned amount",
                        "schema": {
                            "properties": {
                                "error": {"type": "string"}
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}